+++ ksqldb-engine/src/main/java/io/confluent/ksql/planner/LogicalPlanner.java
@@ -65,7 +65,6 @@ import io.confluent.ksql.planner.plan.PlanNodeId;
 import io.confluent.ksql.planner.plan.PreJoinProjectNode;
 import io.confluent.ksql.planner.plan.PreJoinRepartitionNode;
 import io.confluent.ksql.planner.plan.ProjectNode;
-import io.confluent.ksql.planner.plan.RepartitionNode;
 import io.confluent.ksql.planner.plan.SelectionUtil;
 import io.confluent.ksql.planner.plan.SuppressNode;
 import io.confluent.ksql.planner.plan.UserRepartitionNode;
@@ -321,7 +320,7 @@ public class LogicalPlanner {
     return new FilterNode(new PlanNodeId("WhereFilter"), sourcePlanNode, filterExpression);
   }
 
-  private RepartitionNode buildUserRepartitionNode(
+  private UserRepartitionNode buildUserRepartitionNode(
       final PlanNode currentNode,
       final PartitionBy partitionBy
   ) {
@@ -340,7 +339,7 @@ public class LogicalPlanner {
     );
   }
 
-  private RepartitionNode buildInternalRepartitionNode(
+  private PreJoinRepartitionNode buildInternalRepartitionNode(
       final PlanNode source,
       final String side,
       final Expression joinExpression,
