+++ ksqldb-engine/src/main/java/io/confluent/ksql/planner/plan/PreJoinRepartitionNode.java
@@ -15,8 +15,12 @@
 
 package io.confluent.ksql.planner.plan;
 
+import static java.util.Objects.requireNonNull;
+
+import io.confluent.ksql.execution.builder.KsqlQueryBuilder;
 import io.confluent.ksql.execution.expression.tree.Expression;
 import io.confluent.ksql.schema.ksql.LogicalSchema;
+import io.confluent.ksql.structured.SchemaKStream;
 
 /**
  * Node to handle an implicit repartition required to enable a join.
@@ -24,7 +28,10 @@ import io.confluent.ksql.schema.ksql.LogicalSchema;
  * <p>Any join that is not on the key of the stream requires ksql to perform an
  * implicit repartition step before joining.
  */
-public class PreJoinRepartitionNode extends RepartitionNode {
+public class PreJoinRepartitionNode extends SingleSourcePlanNode {
+
+  private final Expression partitionBy;
+  private final LogicalSchema schema;
 
   public PreJoinRepartitionNode(
       final PlanNodeId id,
@@ -32,6 +39,22 @@ public class PreJoinRepartitionNode extends RepartitionNode {
       final LogicalSchema schema,
       final Expression partitionBy
   ) {
-    super(id, source, schema, partitionBy);
+    super(id, source.getNodeOutputType(), source.getSourceName(), source);
+    this.schema = requireNonNull(schema, "schema");
+    this.partitionBy = requireNonNull(partitionBy, "partitionBy");
+  }
+
+  @Override
+  public LogicalSchema getSchema() {
+    return schema;
+  }
+
+  @Override
+  public SchemaKStream<?> buildStream(final KsqlQueryBuilder builder) {
+    return getSource().buildStream(builder)
+        .selectKey(
+            partitionBy,
+            builder.buildNodeContext(getId().toString())
+        );
   }
 }
